V2.0
- Improved build system
- Added stack debug framework
- Removed not used source files
- Added support DMX delta output
- Added Manufacturer PIDs (https://www.gd32-dmx.org/rdm.html)
- Added support for void __libc_init_array(void)
- Changed the layout of the config store.
    - Before upgrading the firmware it is advised to make a backup of the configuration.
    - After the upgrade, then restore the saved configuration.
- lib-artnet : V1.58
- lib-e131 : V1.27
- Removed  lib-artnet4
- Improved Network abstraction layer
- Improved PHY initialization process
- Improved embedded HTTP server
- Added support for secondary (permanent) IP address
    - The node is always available on the secondary IP address (2.x.x.x) for ping and http requests.

+++ lib-c/src/malloc.c
@@ -8,7 +8,7 @@
  * Copyright (C) 2014-2016  R. Stange <rsta2@o2online.de>
  * https://github.com/rsta2/circle/blob/master/lib/alloc.cpp
  */
-/* Copyright (C) 2021 by Arjan van Vught mailto:info@gd32-dmx.org
+/* Copyright (C) 2021-2023 by Arjan van Vught mailto:info@gd32-dmx.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -203,7 +203,7 @@ void *calloc(size_t n, size_t size) {
 		*dst8++ = (uint8_t) 0;
 	}
 
-	assert(((void *)dst8 - (void *)p) == (n * size));
+	assert((size_t)((void *)dst8 - (void *)p) == (n * size));
 
 	return (void *) p;
 }
@@ -250,7 +250,7 @@ void *realloc(void *ptr, size_t size) {
 			*dst8++ = *src8++;
 		}
 
-		assert(((void *)dst8 - (void *)newblk) == size);
+		assert((size_t)((void *)dst8 - (void *)newblk) == size);
 
 		free(ptr);
 	}
